ZSH-compatibility for the IPv6 launch command
The IPv6-compatible launch command in the README.md uses square brackets. Square brackets are fine in `bash`, but they do have a special meaning in `zsh`. However, there is an easy way around it: wrap the URL in "". This doesn't harm `bash` but also makes the command copy&pastable in `zsh`.